Changes structure and dispatch mechanism
The interpreter is about to get a debugger. That makes the interpreter
code to complex for a single file, and therefore I decided to have the
interpreter as separate crate.

I also changed the dispatch mechanism. This way it's a) faster (I hope)
and b) it will be easier to implement a debugger