Finishes data type remodeling
Now you can only get refs from data values that are actually contained.
That breaks the int_lit.les testcase because now the vm tries to get a
&str from from a Data-value holding an int.

The reason is that casting should happen within the language.
And the compiler is responsible to have the correct types on the stack.

so now i need to introduce type checking, type inference and implicit
casting, and then everything will work