Changes everything
Refs use pointers now. Memory has a heap.
A lot of stuff is really unclean now, and I must clean up code and doc.
Tests pass though. Except for the int_lit.les test. Because currently it
is assumed that there is type-checking, it's not however. The
int_lit.les thing will start to work again, once I have type checking
and implicit casting in place