Refactors the AST, and begins introduction of Functions
The main thing in this Commit is the AST-Remodel, the introduction of
the FnFragment, as well as type sets, are an accident, because I forgot
to commit the AST first. So I tried to get it to compile and pass all
tests, but there are a lot of lose ends in this commit now.